# api/openapi.yaml
openapi: 3.0.3
info:
  title: Velox Billing API
  description: |
    Open-source usage-based billing engine. Handles pricing, subscriptions,
    usage metering, invoice generation, Stripe payments, dunning, and credits.
  version: "2026-04-07"
  license:
    name: MIT
    url: https://github.com/sagarsuperuser/velox/blob/main/LICENSE
  contact:
    name: Velox
    url: https://github.com/sagarsuperuser/velox
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication. Use `Authorization: Bearer vlx_secret_...`
        Key types: vlx_platform_ (tenant mgmt), vlx_secret_ (full access), vlx_pub_ (restricted)
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: string
          example: vlx_cus_42014438d82ddf33
        external_id:
          type: string
        display_name:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [active, archived]
        created_at:
          type: string
          format: date-time
...
paths:
...
  /v1/customers:
    post:
      tags: [Customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [external_id, display_name]
              properties:
                external_id:
                  type: string
                display_name:
                  type: string
                email:
                  type: string
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "409":
          description: Customer already exists
    get:
      tags: [Customers]
      summary: List customers
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Customer list
  /v1/customers/{id}/rotate-cost-dashboard-token:
    post:
      tags: [Customers]
      summary: Rotate the public cost-dashboard token (ADR-031)
      description: |
        Mints a fresh `vlx_pcd_…` token and writes it to the customer
        row. The previous token is invalidated immediately (read-only
        public surface; no grace window). The plaintext token is the
        sole credential for `GET /v1/public/cost-dashboard/{token}`
        and is shown ONCE — Velox never returns it again after this
        response. Audit log records the rotation without the token.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rotated
          content:
            application/json:
              schema:
                type: object
                required: [token, public_url]
                properties:
                  token:
                    type: string
                    example: vlx_pcd_a3f9c4...
                  public_url:
                    type: string
                    example: https://api.velox.dev/v1/public/cost-dashboard/vlx_pcd_a3f9c4...
        "404":
          description: Customer not found
  /v1/public/cost-dashboard/{token}:
    get:
      tags: [Public]
      summary: Public cost-dashboard projection (ADR-031)
      description: |
        Unauthenticated. The `vlx_pcd_…` token in the URL is the sole
        credential — partners embed via iframe / fetch from their own
        app without an API key. Sanitized projection: customer_id,
        tenant_id, billing_period, subscriptions, usage[], totals,
        projected_total_cents. PII (email, display_name, external_id,
        metadata, billing_profile) is NEVER on this response. Wrong
        prefix or unknown token → 401 (anti-enumeration). No active
        sub → 200 with `billing_period.source = "no_subscription"`
        and empty arrays.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Projection
        "401":
          description: Invalid token
        "429":
          description: Rate limited (60/min/IP)
...